+++ plugins/trivy/pkg/image/newest_images.go
@@ -4,14 +4,16 @@ import (
 	"context"
 	"strings"
 
-	"github.com/docker/docker/api/types"
-	"github.com/genuinetools/reg/registry"
-	"github.com/genuinetools/reg/repoutils"
-	version "github.com/mcuadros/go-version"
+	"github.com/google/go-containerregistry/pkg/authn"
+	"github.com/google/go-containerregistry/pkg/name"
+	"github.com/google/go-containerregistry/pkg/v1/remote"
+
+	semver "github.com/Masterminds/semver/v3"
+
 	"github.com/sirupsen/logrus"
 )
 
-var specific = []string{
+var knownPreReleases = []string{
 	"centos-7",
 	"debian-8",
 	"debian-9",
@@ -28,13 +30,16 @@ const MaxNewestVersionsToScan = 1
 // GetNewestVersions returns newest versions and newest version within same major version
 func GetNewestVersions(ctx context.Context, repo, tag string) ([]string, error) {
 	logrus.Info("Started retrieving newest versions for ", repo, ":", tag)
-	tags, err := fetchTags(ctx, repo, tag)
+	tags, err := fetchTags(ctx, repo)
 	if err != nil {
 		logrus.Error("Error fetching tags for ", repo, ":", tag, err)
 		return nil, err
 
 	}
-	newest := filterAndSort(tags, tag)
+	newest, err := filterAndSort(tags, tag)
+	if err != nil {
+		return nil, err
+	}
 	logrus.Info("Finished retrieving newest versions for ", repo, ":", tag)
 	if len(newest) <= MaxNewestVersionsToScan {
 		return newest, nil
@@ -42,52 +47,58 @@ func GetNewestVersions(ctx context.Context, repo, tag string) ([]string, error) 
 	return newest[len(newest)-MaxNewestVersionsToScan:], nil
 }
 
-func fetchTags(ctx context.Context, imageName, tag string) ([]string, error) {
-	image, err := registry.ParseImage(imageName)
+func fetchTags(ctx context.Context, imageRepoName string) ([]string, error) {
+	repository, err := name.NewRepository(imageRepoName)
 	if err != nil {
 		return nil, err
 	}
-	r, err := createRegistryClient(ctx, image.Domain)
-	if err != nil {
-		return nil, err
-	}
-	tags, err := r.Tags(ctx, image.Path)
+	tags, err := remote.List(repository, remote.WithAuthFromKeychain(authn.DefaultKeychain), remote.WithContext(ctx))
 	if err != nil {
 		return nil, err
 	}
 	return tags, nil
 }
 
-func createRegistryClient(ctx context.Context, domain string) (*registry.Registry, error) {
-	auth := types.AuthConfig{}
-	if domain == "docker.io" {
-		auth.ServerAddress = repoutils.DefaultDockerRegistry
+func filterAndSort(suggestedTags []string, curTagStr string) ([]string, error) {
+	curVersion, err := semver.NewVersion(curTagStr)
+	if err != nil {
+		logrus.Infof("could not parse current tag semver %s: %v", curTagStr, err)
+		return []string{}, nil
 	}
-	return registry.New(ctx, auth, registry.Opt{
-		Domain:   domain,
-		Insecure: false,
-		Debug:    false,
-		SkipPing: false,
-		NonSSL:   false,
-	})
-}
 
-func filterAndSort(tags []string, currentTag string) []string {
-	newest := []string{}
-	c := version.NewConstrainGroupFromString(">" + currentTag)
-	currentTagSpecificToken := GetSpecificToken(currentTag)
-	for _, targetTag := range tags {
-		targetTagSpecificToken := GetSpecificToken(targetTag)
-		if c.Match(targetTag) && currentTagSpecificToken == targetTagSpecificToken {
-			newest = append(newest, targetTag)
+	constraint, err := semver.NewConstraint(">" + curTagStr)
+	if err != nil {
+		return nil, err
+	}
+
+	newest := []*semver.Version{}
+	for _, tag := range suggestedTags {
+		v, err := semver.NewVersion(tag)
+		if err != nil {
+			logrus.Infof("could not parse tag semver %s: %v", tag, err)
+			continue
+		}
+
+		cPre := curVersion.Prerelease()
+		vPre := v.Prerelease()
+		if cPre != "" && vPre != "" && cPre != vPre {
+			logrus.Infof("pre-releases does not match: %s != %s", cPre, vPre)
+			continue
+		}
+
+		check, errors := constraint.Validate(v)
+		if check {
+			newest = append(newest, v)
+		} else {
+			logrus.Debug(errors) // for debugging only
 		}
 	}
-	version.Sort(newest)
-	return newest
+	Sort(newest)
+	return Versions(newest).ToStringSlice(), nil
 }
 
 func GetSpecificToken(tag string) string {
-	for _, v := range specific {
+	for _, v := range knownPreReleases {
 		if strings.Contains(tag, v) {
 			return v
 		}