+++ plugins/trivy/CHANGELOG.md
@@ -2,6 +2,15 @@
 ## 0.20.0
 * Offline support
 
+## 0.20.2
+* improve image tags recommendations
+
+## 0.20.1
+* update to go 1.18 and update packages
+
+## 0.20.0
+* Fixing failing tags fetch from `quay.io`
+
 ## 0.19.3
 * Update alpine to remove CVE